+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/bookmark/BookmarksFragment.kt
@@ -200,7 +200,7 @@ class BookmarksFragment : BaseFragment() {
     }
 
     private fun onShareBookmarkClick() {
-        bookmarksViewModel.onShareClicked(requireContext())
+        bookmarksViewModel.onShareClicked(parentFragmentManager)
     }
 
     private fun onEditBookmarkClick() {